refactoring - marriage rings, etc.
git-svn-id: https://svn.code.sf.net/p/genj/code/trunk@852 fbd980a2-0dff-11df-b4b7-0b05b74d0985

+++ genj/src/core/genj/tree/TreeViewSettings.java
@@ -140,11 +140,11 @@ public class TreeViewSettings extends JTabbedPane implements Settings {
     colors.apply(); //FIXME we shouldn't have to call repaint
     // metrics
     view.getModel().setMetrics(new TreeMetrics(
-      sliderCmIndiWidth .getValue(),
-      sliderCmIndiHeight.getValue(),
-      sliderCmFamWidth  .getValue(),
-      sliderCmFamHeight .getValue(),
-      sliderCmPadding   .getValue()
+      (float)sliderCmIndiWidth .getValue(),
+      (float)sliderCmIndiHeight.getValue(),
+      (float)sliderCmFamWidth  .getValue(),
+      (float)sliderCmFamHeight .getValue(),
+      (float)sliderCmPadding   .getValue()
     ));
     // blueprints
     view.setBlueprints(blueprintList.getSelection());
